+++ tests/test_opd_data2.py
@@ -220,7 +220,6 @@ def test_agency_filter():
 		assert len(table.table)==100
 		assert table.table[table.agency_field].nunique()==1
 		assert table.table.iloc[0][table.agency_field] == agency
-pytest -k test_save_load test_opd_data2.py
 
 @pytest.mark.parametrize('save,fname,load',[('to_csv','get_csv_filename'),
 											('to_csv','get_csv_filename','load_csv'),